refactor: banner to be separate component
